Fixed testing errors with help of Greg and Earl. Changed architecture a bit from their comments as well (moved stuff in response.js to router.js).
