+++ lib/router.js
@@ -1,6 +1,6 @@
 'use strict';
 
-require(__dirname + '/response');
+var fs = require('fs');
 
 var routes = {
 'GET': {},
@@ -19,7 +19,6 @@ module.exports = exports = function() {
 
 Router.prototype.get = function(route, callback) {
   routes.GET[route] = function(req, res) {
-    res.status = 200;
     callback(req, res);
     res.end();
   };
@@ -27,7 +26,6 @@ Router.prototype.get = function(route, callback) {
 
 Router.prototype.post = function(route, callback) {
   routes.POST[route] = function(req, res) {
-    res.status = 200;
     callback(req, res);
     res.end();
   };
@@ -47,6 +45,40 @@ Router.prototype.delete = function(route, callback) {
 
 Router.prototype.route = function(req, res) {
   if(typeof routes[req.method][req.url] === 'function') {
+    if(!res.set) {
+      res.set = function() {
+        this.setHeader(arguments[0], arguments[1]);
+      };
+    }
+
+    if(!res.send) {
+      res.send = function(body) {
+        this.write(body);
+      };
+    }
+
+    if(!res.json) {
+      res.json = function(req) {
+        this.writeHead(200, {'Content-Type': 'application/json'});
+
+        var body = '';
+        req.on('data', function(chunk) {
+          body += chunk;
+        });
+
+        req.on('end', function() {
+          if(!fs.existsSync(__dirname + '/../data')) {
+            fs.mkdirSync(__dirname + '/../data');
+          }
+
+          var filesArr = fs.readdirSync(__dirname + '/../data');
+          fs.writeFileSync(__dirname + '/../data/request' + filesArr.length + '.json', body);
+        });
+
+        this.send(JSON.stringify({msg: 'json stored'}));
+      };
+    }
+
     routes[req.method][req.url](req, res);
   }
 };
